feat(frontend): move link to tab3 from work and school sections to service stats sections
