+++ frontend/src/components/sections/Coverage.tsx
@@ -1,13 +1,16 @@
-import { useSearchParams } from 'react-router';
+import { useLocation, useSearchParams } from 'react-router';
 import { flatSectionBundleIds } from '.';
 import { useAppData, useSectionsVisibility, useToggleSectionItemVisibility } from '../../hooks';
 import { notEmpty, shouldRenderStatistic } from '../../utils';
-import { Section, Statistic } from '../common';
+import { Button, Section, SectionEntry, Statistic } from '../common';
+import { StatisticContainer } from '../common/Statistic/StatisticContainer';
+import { TAB_3_FRAGMENT } from '../navigation';
 import { useComparisonModeState } from '../options';
 
 export function Coverage() {
   const { scenarios } = useAppData();
   const { data: scenariosData } = scenarios;
+  const { search } = useLocation();
 
   const [isComparing] = useComparisonModeState();
   const [searchParams] = useSearchParams();
@@ -19,7 +22,7 @@ export function Coverage() {
     selectedRouteIds.includes(future.__routeId)
   );
 
-  const [visibleSections] = useSectionsVisibility();
+  const [visibleSections, , visibleTabs] = useSectionsVisibility();
   const { editMode, handleClick } = useToggleSectionItemVisibility('Future.Coverage');
   const shouldRender = shouldRenderStatistic.bind(
     null,
@@ -27,6 +30,20 @@ export function Coverage() {
     flatSectionBundleIds['Future.Coverage'],
     editMode
   );
+
+  const jobAccessSearch = (() => {
+    const currentSearchParams = new URLSearchParams(search);
+
+    const selectedAreas = currentSearchParams.get('areas')?.split(',').filter(notEmpty) || [];
+    const selectedSeasons = currentSearchParams.get('seasons')?.split(',').filter(notEmpty) || [];
+
+    const selectedSeasonAreas = selectedAreas.flatMap((area) => {
+      return selectedSeasons.map((season) => `${area}::${season}`);
+    });
+
+    currentSearchParams.set('jobAreas', selectedSeasonAreas.join(','));
+    return currentSearchParams.toString() ? `?${currentSearchParams.toString()}` : '';
+  })();
 
   return (
     <Section title="Coverage">
@@ -74,6 +91,23 @@ export function Coverage() {
         unit="square miles"
         onClick={handleClick('cov')}
       />
+      {shouldRender('jadl') && (!visibleTabs || visibleTabs.includes(TAB_3_FRAGMENT)) ? (
+        <SectionEntry f={{ gridColumn: 'span 2' }}>
+          <StatisticContainer
+            onClick={handleClick('jadl')}
+            style={{ opacity: shouldRender('jadl') === 'partial' ? 0.5 : 1 }}
+          >
+            <div>
+              <div style={{ fontSize: '0.875rem' }}>
+                What sectors could be served by this new route?
+              </div>
+              <Button href={'#/job-access' + jobAccessSearch}>
+                Explore industry/sector of employment
+              </Button>
+            </div>
+          </StatisticContainer>
+        </SectionEntry>
+      ) : null}
     </Section>
   );
 }
